chore: moves snaps specs to tests
