+++ tests/smoke/snaps/test-snap-installed.spec.ts
@@ -1,9 +1,9 @@
-import FixtureBuilder from '../../../tests/framework/fixtures/FixtureBuilder';
-import { withFixtures } from '../../../tests/framework/fixtures/FixtureHelper';
-import { FlaskBuildTests } from '../../tags';
+import FixtureBuilder from '../../framework/fixtures/FixtureBuilder';
+import { withFixtures } from '../../framework/fixtures/FixtureHelper';
+import { FlaskBuildTests } from '../../../e2e/tags';
 
-import TestSnaps from '../../pages/Browser/TestSnaps';
-import { loginToApp, navigateToBrowserView } from '../../viewHelper';
+import TestSnaps from '../../../e2e/pages/Browser/TestSnaps';
+import { loginToApp, navigateToBrowserView } from '../../../e2e/viewHelper';
 
 jest.setTimeout(150_000);
