+++ tests/smoke/snaps/test-snap-preinstalled.spec.ts
@@ -1,11 +1,11 @@
-import { FlaskBuildTests } from '../../tags';
-import { loginToApp, navigateToBrowserView } from '../../viewHelper';
-import FixtureBuilder from '../../../tests/framework/fixtures/FixtureBuilder';
-import { withFixtures } from '../../../tests/framework/fixtures/FixtureHelper';
-import TestSnaps from '../../pages/Browser/TestSnaps';
-import Assertions from '../../../tests/framework/Assertions';
-import { getEventsPayloads } from '../../../tests/helpers/analytics/helpers';
-import TestHelpers from '../../helpers';
+import { FlaskBuildTests } from '../../../e2e/tags';
+import { loginToApp, navigateToBrowserView } from '../../../e2e/viewHelper';
+import FixtureBuilder from '../../framework/fixtures/FixtureBuilder';
+import { withFixtures } from '../../framework/fixtures/FixtureHelper';
+import TestSnaps from '../../../e2e/pages/Browser/TestSnaps';
+import Assertions from '../../framework/Assertions';
+import { getEventsPayloads } from '../../helpers/analytics/helpers';
+import TestHelpers from '../../../e2e/helpers';
 
 jest.setTimeout(150_000);
