+++ tests/smoke/snaps/test-snap-rpc.spec.ts
@@ -1,8 +1,8 @@
-import FixtureBuilder from '../../../tests/framework/fixtures/FixtureBuilder';
-import TestSnaps from '../../pages/Browser/TestSnaps';
-import { FlaskBuildTests } from '../../tags';
-import { loginToApp, navigateToBrowserView } from '../../viewHelper';
-import { withFixtures } from '../../../tests/framework/fixtures/FixtureHelper';
+import FixtureBuilder from '../../framework/fixtures/FixtureBuilder';
+import TestSnaps from '../../../e2e/pages/Browser/TestSnaps';
+import { FlaskBuildTests } from '../../../e2e/tags';
+import { loginToApp, navigateToBrowserView } from '../../../e2e/viewHelper';
+import { withFixtures } from '../../framework/fixtures/FixtureHelper';
 
 jest.setTimeout(150_000);
